refactor: reactor apply command
