+++ src/spacectl/modules/__init__.py
@@ -1,9 +1,8 @@
-from spacectl.modules.resource.resource_task import ResourceTask
-from spacectl.modules.shell.shell_task import ShellTask
-from spacectl.modules.export_google_sheets.export_google_sheets_task import ExportGoogleSheetsTask
+from spacectl.modules import shell, print, resource, export_google_sheets
 
 MODULES = {
-    'resource': ResourceTask,
-    'shell': ShellTask,
-    'export-google-sheets': ExportGoogleSheetsTask
+    'shell': shell.Task,
+    'print': print.Task,
+    'resource': resource.Task,
+    'export-google-sheets': export_google_sheets.Task
 }
