--- src/spacectl/modules/shell/__init__.py
@@ -1,3 +1,0 @@
-
-
-
